🧪 test: Test and adjusting breed route

+++ app/routes/breed.py
@@ -1,19 +1,29 @@
 from http import HTTPStatus
 
-from fastapi import APIRouter
+from fastapi import APIRouter, HTTPException
 from sqlalchemy import select
-from sqlalchemy import update as up
 
 from app.models import Breed
 from app.schemas import Message
-from app.schemas.breed import BreedList, BreedSchema
-from app.utils import T_Session
+from app.schemas.breed import BreedCreateUpdate, BreedList, BreedSchema
+from app.utils import T_CurrentUser, T_Session
 
 router = APIRouter(prefix='/animal/breed', tags=['Animal Breed'])
 
 
 @router.post('/', status_code=HTTPStatus.CREATED, response_model=BreedSchema)
-async def create(schema: BreedSchema, session: T_Session):
+async def create(
+    schema: BreedCreateUpdate, session: T_Session, current_user: T_CurrentUser
+):
+    db_breed = await session.scalar(
+        select(Breed).where(Breed.name == schema.name)
+    )
+
+    if db_breed:
+        raise HTTPException(
+            status_code=HTTPStatus.CONFLICT, detail='Breed already exists'
+        )
+
     db_breed = Breed(**schema.model_dump())
 
     session.add(db_breed)
@@ -23,31 +33,43 @@ async def create(schema: BreedSchema, session: T_Session):
     return db_breed
 
 
-@router.get('/{breed_id}', response_model=BreedSchema)
-async def get_by_id(breed_id: int, session: T_Session):
-    db_breed = await session.scalar(select(Breed).where(Breed.id == breed_id))
-
-    return db_breed
-
-
 @router.get('/', response_model=BreedList)
-async def get_all(session: T_Session):
+async def get_all(session: T_Session, current_user: T_CurrentUser):
     db_breed = await session.scalars(select(Breed))
 
     return {'breeds': db_breed.all()}
 
 
-@router.patch('/{breed_id}', response_model=BreedSchema)
-async def update(breed_id: int, schema: BreedSchema, session: T_Session):
-    query = up(Breed).where(Breed.id == breed_id).values(**schema.model_dump())
-    db_breed = await session.scalar(query.returning(Breed))
+@router.patch('/{name}', response_model=BreedSchema)
+async def update(
+    name: str,
+    schema: BreedCreateUpdate,
+    session: T_Session,
+    current_user: T_CurrentUser,
+):
+    db_breed = await session.scalar(select(Breed).where(Breed.name == name))
 
+    if not db_breed:
+        raise HTTPException(
+            status_code=HTTPStatus.NOT_FOUND, detail='Breed not found'
+        )
+
+    db_breed.name = schema.name
+
+    session.add(db_breed)
+    await session.commit()
+    await session.refresh(db_breed)
     return db_breed
 
 
-@router.delete('/{breed_id}', response_model=Message)
-async def delete(breed_id: int, session: T_Session):
-    db_breed = session.scalar(select(Breed).where(Breed.id == breed_id))
+@router.delete('/{name}', response_model=Message)
+async def delete(name: str, session: T_Session, current_user: T_CurrentUser):
+    db_breed = await session.scalar(select(Breed).where(Breed.name == name))
+
+    if not db_breed:
+        raise HTTPException(
+            status_code=HTTPStatus.NOT_FOUND, detail='Breed not found'
+        )
 
     await session.delete(db_breed)
     await session.commit()
